# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DrépaConnect API — plateforme de lutte contre la drépanocytose
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # PATIENTS
...
  /crises:
    get:
      operationId: listCrises
      tags: [crises]
      summary: List crisis entries for current patient
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Crisis list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Crisis"
    post:
      operationId: createCrisis
      tags: [crises]
      summary: Log a crisis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrisisInput"
      responses:
        "201":
          description: Crisis logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crisis"
  /crises/{id}:
    get:
      operationId: getCrisis
      tags: [crises]
      summary: Get crisis entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crisis entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crisis"
    patch:
      operationId: updateCrisis
      tags: [crises]
      summary: Update crisis entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrisisUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crisis"
    delete:
      operationId: deleteCrisis
      tags: [crises]
      summary: Delete crisis entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /crises/stats/summary:
    get:
      operationId: getCrisisStats
      tags: [crises]
      summary: Crisis statistics summary
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Crisis statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrisisStats"

  # MEDICATIONS
...
  /centers:
    get:
      operationId: listCenters
      tags: [centers]
      summary: List screening centers
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: province
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Centers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreeningCenter"
    post:
      operationId: createCenter
      tags: [centers]
      summary: Create screening center (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningCenterInput"
      responses:
        "201":
          description: Center created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningCenter"
  /centers/{id}:
    get:
      operationId: getCenter
      tags: [centers]
      summary: Get screening center
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Center details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningCenter"

  # GENETIC CALCULATOR
...
  /blood-requests:
    get:
      operationId: listBloodRequests
      tags: [blood-requests]
      summary: List active blood requests
      parameters:
        - name: bloodType
          in: query
          schema:
            type: string
        - name: urgent
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodRequest"
    post:
      operationId: createBloodRequest
      tags: [blood-requests]
      summary: Create blood donation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"
  /blood-requests/{id}/respond:
    post:
      operationId: respondToBloodRequest
      tags: [blood-requests]
      summary: Donor responds to blood request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodResponseInput"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

  # PREGNANCY
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [PATIENT, DOCTOR, DONOR, HELPER, ADMIN]
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [PATIENT, DOCTOR, DONOR, HELPER, ADMIN]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string

    # PATIENT
...
    Crisis:
      type: object
      required: [id, patientId, intensity, startedAt, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        intensity:
          type: integer
          minimum: 1
          maximum: 10
        location:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        triggers:
          type: ["string", "null"]
        treatment:
          type: ["string", "null"]
        hospitalVisit:
          type: boolean
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CrisisInput:
      type: object
      required: [patientId, intensity, startedAt]
      properties:
        patientId:
          type: integer
        intensity:
          type: integer
          minimum: 1
          maximum: 10
        location:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        triggers:
          type: ["string", "null"]
        treatment:
          type: ["string", "null"]
        hospitalVisit:
          type: boolean
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CrisisUpdate:
      type: object
      properties:
        intensity:
          type: integer
          minimum: 1
          maximum: 10
        location:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        triggers:
          type: ["string", "null"]
        treatment:
          type: ["string", "null"]
        hospitalVisit:
          type: boolean
        endedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CrisisStats:
      type: object
      required: [total, avgIntensity, hospitalVisits, lastCrisisDate]
      properties:
        total:
          type: integer
        avgIntensity:
          type: number
        hospitalVisits:
          type: integer
        lastCrisisDate:
          type: ["string", "null"]
        thisMonth:
          type: integer
        thisYear:
          type: integer
...
    ScreeningCenter:
      type: object
      required: [id, name, city, province, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        province:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        services:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    ScreeningCenterInput:
      type: object
      required: [name, city, province]
      properties:
        name:
          type: string
        city:
          type: string
        province:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        services:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]

    # GENETIC CALCULATOR
...
    BloodRequest:
      type: object
      required: [id, bloodType, city, urgent, status, createdAt]
      properties:
        id:
          type: integer
        requestedBy:
          type: ["integer", "null"]
        patientName:
          type: string
        bloodType:
          type: string
          enum: [A_POS, A_NEG, B_POS, B_NEG, AB_POS, AB_NEG, O_POS, O_NEG]
        city:
          type: string
        hospital:
          type: ["string", "null"]
        urgent:
          type: boolean
        status:
          type: string
          enum: [OPEN, FULFILLED, CANCELLED]
        message:
          type: ["string", "null"]
        respondedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
    BloodRequestInput:
      type: object
      required: [patientName, bloodType, city]
      properties:
        patientName:
          type: string
        bloodType:
          type: string
          enum: [A_POS, A_NEG, B_POS, B_NEG, AB_POS, AB_NEG, O_POS, O_NEG]
        city:
          type: string
        hospital:
          type: ["string", "null"]
        urgent:
          type: boolean
        message:
          type: ["string", "null"]
    BloodResponseInput:
      type: object
      required: [donorId]
      properties:
        donorId:
          type: integer
        message:
          type: ["string", "null"]

    # PREGNANCY
...